🏗️(./store): rename to ./stores
