+++ src/pages/FamiliarizationPage.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import { _ } from 'svelte-i18n'
     import AlphanumericInput from '../components/AlphanumericInput/AlphanumericInput.svelte'
-    import { swiper } from '../store'
+    import { swiper } from '../stores'
 
     const CHUNK_LENGTH = 9
     const SEQUENCE = 'T890SOLA6 RF6628GB8'
