+++ src/stores/swiper.ts
